DeleteRoom e deletePerson atualizadas e funcionando

+++ src/functions/people/deletePerson.ts
@@ -1,6 +1,7 @@
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
 import { Handler } from "src/errors/Handler";
 import PeopleRepositories from "src/repositories/implementations/PeopleRepositories";
+import RoomsRepositories from "src/repositories/implementations/RoomsRepositories";
 import NotFoundError from "src/errors/NotFoundError";
 import { ok } from "src/utils/Returns";
 
@@ -14,6 +15,13 @@ const deletePerson = async (
         throw new NotFoundError("Pessoa não encontrada!");
 
     const database = new PeopleRepositories();
+    const person = await database.findByEmail(email);
+    if (person === undefined)
+        throw new NotFoundError("Pessoa não encontrada!");
+
+    const database2 = new RoomsRepositories();
+    const room = await database2.findById (person.id_quarto);
+    await database2.removePerson (room, email);
 
     database.delete(email);
