+++ src/functions/rooms/deleteRoom.ts
@@ -1,7 +1,9 @@
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
 import { Handler } from "src/errors/Handler";
 import RoomsRepositories from "src/repositories/implementations/RoomsRepositories";
+import PeopleRepositories from "src/repositories/implementations/PeopleRepositories";
 import NotFoundError from "src/errors/NotFoundError";
+import ClientError from "src/errors/ClientError";
 import { ok } from "src/utils/Returns";
 
 
@@ -9,16 +11,25 @@ const deleteRoom = async (
     event: APIGatewayProxyEvent
   ): Promise<APIGatewayProxyResult> => {
   
-    const { id } = JSON.parse(event.body);
+    const { id } = event.pathParameters;
     if (id === undefined)
-        throw new NotFoundError("Quarto não encontrado!");
+        throw new ClientError("Quarto não formatado!");
 
     const database = new RoomsRepositories();
     const room = await database.findById(id);
+    if (room === undefined)
+        throw new NotFoundError("Quarto não encontrado!");
 
+    const database2 = new PeopleRepositories();
+    const listaPessoas = await database.listPeopleInsideRoom(room);
+    for (let i = 0; i < listaPessoas.length; i++) {
+        const person = await database2.findByEmail(listaPessoas[i]);
+        await database2.removeRoom(person);
+    }
+    
     database.delete(room);
 
-    return ok("Cama deletada com sucesso", "message");
+    return ok("Quarto deletado com sucesso", "message");
   };
 
   export const handler = Handler(deleteRoom);
